feat(agent): donut for portfolio breakdown, extract shared InteractiveDonut
Pulled the segmented d3.arc donut out of TopCategoriesCard into a shared
InteractiveDonut at components/InteractiveDonut.tsx. Same visuals (slice
gap, rounded corners, hover-bump, dim non-hovered, pointer-type-aware
click handling for touch), now driven by props: size, strokeWidth,
centerLabel, pctSuffix.

Two consumers:

- TopCategoriesCard imports it back in at the same 220px size, passes
  pctSuffix="of spending" so the center hover text stays unchanged
  ("47% of spending"). Visually identical to before — verified by
  reading the path generation against the prior inline implementation.

- PortfolioBreakdownWidget swaps its stacked bar + percent legend for
  the donut at 160px with pctSuffix="of portfolio". The widget was
  shorter than the line chart next to it, so the asset-class section
  felt visually orphaned in the two-column pair layout; the donut
  fills that vertical space and matches the dashboard's spending
  donut treatment exactly. Tight color-coded legend stays below the
  donut and cross-highlights on hover, same pattern as the dashboard.

Future categorical breakdowns (income by source, accounts by type,
whatever) can drop in the same InteractiveDonut.
